requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
filelock==3.18.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
fsspec==2025.7.0
greenlet==3.2.3
huggingface-hub==0.33.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
langdetect==1.0.9
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pandas==2.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
sympy==1.14.0
tokenizers==0.21.2
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
gunicorn
azure-cognitiveservices-speech==1.40.0
requests==2.31.0
redis==5.0.8